#80 [MOVE] 외부 api 파일이동
