+++ src/apis/external/useDeleteExternals.ts
@@ -1,7 +1,26 @@
 import { useMutation } from '@tanstack/react-query';
-import { deleteExternalItem } from './external';
 import queryClient from '../../utils/queryClient';
 import { queryKey } from '../../constants/queryKey';
+import type { RequestExternalListDto } from '../../types/external';
+import type { CommonResponse } from '../../types/common';
+import { axiosInstance } from '../axios';
+
+// external/externalHome.tsx
+export const deleteExternalItem = async ({
+  teamId,
+  externalIds,
+}: RequestExternalListDto): Promise<CommonResponse> => {
+  try {
+    const { data } = await axiosInstance.delete(`/api/teams/${teamId}/externals`, {
+      data: externalIds,
+    });
+
+    return data;
+  } catch (error) {
+    console.error('Error deleting external item:', error);
+    throw error;
+  }
+};
 
 export const useDeleteExternals = () => {
   return useMutation({
